+++ addOns/bruteforce/src/main/java/com/sittinglittleduck/DirBuster/workGenerators/WorkerGeneratorURLFuzz.java
@@ -23,11 +23,11 @@
 package com.sittinglittleduck.DirBuster.workGenerators;
 
 import com.sittinglittleduck.DirBuster.BaseCase;
-import com.sittinglittleduck.DirBuster.Config;
 import com.sittinglittleduck.DirBuster.DirToCheck;
 import com.sittinglittleduck.DirBuster.GenBaseCase;
-import com.sittinglittleduck.DirBuster.HTTPHeader;
+import com.sittinglittleduck.DirBuster.HttpStatus;
 import com.sittinglittleduck.DirBuster.Manager;
+import com.sittinglittleduck.DirBuster.SimpleHttpClient.HttpMethod;
 import com.sittinglittleduck.DirBuster.WorkUnit;
 import java.io.BufferedReader;
 import java.io.FileInputStream;
@@ -39,8 +39,6 @@ import java.net.URL;
 import java.net.URLEncoder;
 import java.util.Vector;
 import java.util.concurrent.BlockingQueue;
-import org.apache.commons.httpclient.HttpClient;
-import org.apache.commons.httpclient.methods.HeadMethod;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -60,9 +58,6 @@ public class WorkerGeneratorURLFuzz implements Runnable {
     // find bug UuF
     // private String failString = "thereIsNoWayThat-You-CanBeThere";
     // private HttpURLConnection urlConn;
-    // find bug UuF
-    // HttpState initialState;
-    HttpClient httpclient;
 
     private String urlFuzzStart;
     private String urlFuzzEnd;
@@ -90,8 +85,6 @@ public class WorkerGeneratorURLFuzz implements Runnable {
         inputFile = manager.getInputFile();
         firstPart = manager.getFirstPartOfURL();
 
-        httpclient = manager.getHttpclient();
-
         urlFuzzStart = manager.getUrlFuzzStart();
         urlFuzzEnd = manager.getUrlFuzzEnd();
     }
@@ -107,7 +100,6 @@ public class WorkerGeneratorURLFuzz implements Runnable {
         try {
             manager.setURLFuzzGenFinished(false);
             String currentDir = "/";
-            int failcode = 404;
             String line;
             Vector extToCheck = new Vector(10, 5);
             boolean recursive = true;
@@ -131,16 +123,14 @@ public class WorkerGeneratorURLFuzz implements Runnable {
             if (manager.getAuto()) {
                 try {
                     URL headurl = new URL(firstPart);
-                    HeadMethod httphead = new HeadMethod(headurl.toString());
-                    Vector HTTPheaders = manager.getHTTPHeaders();
-                    for (int a = 0; a < HTTPheaders.size(); a++) {
-                        HTTPHeader httpHeader = (HTTPHeader) HTTPheaders.elementAt(a);
-                        httphead.setRequestHeader(httpHeader.getHeader(), httpHeader.getValue());
-                    }
-                    httphead.setFollowRedirects(Config.followRedirects);
-                    int responceCode = httpclient.executeMethod(httphead);
+                    int responceCode =
+                            manager.getHttpClient()
+                                    .send(HttpMethod.HEAD, headurl.toString())
+                                    .getStatusCode();
                     LOG.debug("Response code for head check = {}", responceCode);
-                    if (responceCode == 501 || responceCode == 400 || responceCode == 405) {
+                    if (responceCode == HttpStatus.NOT_IMPLEMENTED
+                            || responceCode == HttpStatus.BAD_REQUEST
+                            || responceCode == HttpStatus.METHOD_NOT_ALLOWED) {
                         LOG.debug(
                                 "Changing to GET only HEAD test returned 501(method no implmented) or a 400");
                         manager.setAuto(false);
@@ -167,13 +157,13 @@ public class WorkerGeneratorURLFuzz implements Runnable {
                 }
 
                 if (!line.startsWith("#")) {
-                    String method;
+                    HttpMethod method;
                     if (manager.getAuto()
                             && !baseCaseObj.useContentAnalysisMode()
                             && !baseCaseObj.isUseRegexInstead()) {
-                        method = "HEAD";
+                        method = HttpMethod.HEAD;
                     } else {
-                        method = "GET";
+                        method = HttpMethod.GET;
                     }
 
                     // url encode all the items